backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.38
asyncpg>=0.30.0
pgvector>=0.3.6
anthropic>=0.45.0
google-genai>=0.2.0
boto3>=1.35.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.12
httpx>=0.28.0
pillow>=11.1.0
open-clip-torch>=2.24.0
rembg>=2.0.50
onnxruntime>=1.16.0
alembic>=1.15.0
pydantic-settings>=2.7.0
email-validator>=2.1.0
python-dotenv>=1.0.1
redis>=5.2.0

# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-httpx>=0.32.0
aiosqlite>=0.20.0
